+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/index/ContentIndex.java
@@ -21,6 +21,7 @@ import com.google.gson.JsonParser;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.action.admin.indices.create.CreateIndexRequest;
+import org.opensearch.OpenSearchTimeoutException;
 import org.opensearch.action.bulk.BulkRequest;
 import org.opensearch.action.bulk.BulkResponse;
 import org.opensearch.action.delete.DeleteRequest;
@@ -35,6 +36,10 @@ import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.xcontent.ToXContent;
 import org.opensearch.core.xcontent.XContentParser;
 import org.opensearch.index.mapper.StrictDynamicMappingException;
+import org.opensearch.index.query.QueryBuilders;
+import org.opensearch.index.reindex.BulkByScrollResponse;
+import org.opensearch.index.reindex.DeleteByQueryAction;
+import org.opensearch.index.reindex.DeleteByQueryRequestBuilder;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
@@ -83,8 +88,8 @@ public class ContentIndex {
     /**
      * This constructor is only used on tests.
      *
-     * @param client @Client (mocked).
-     * @param pluginSettings @PluginSettings (mocked).
+     * @param client Client (mocked).
+     * @param pluginSettings PluginSettings (mocked).
      */
     public ContentIndex(Client client, PluginSettings pluginSettings) {
         this.pluginSettings = pluginSettings;
@@ -327,5 +332,21 @@ public class ContentIndex {
      */
     public boolean exists() {
         return ClusterInfo.indexExists(this.client, ContentIndex.INDEX_NAME);
+    }
+
+
+    /** Clears all documents from the {@link ContentIndex#INDEX_NAME} index. */
+    public void clear() {
+        try {
+            DeleteByQueryRequestBuilder deleteByQuery =
+                    new DeleteByQueryRequestBuilder(this.client, DeleteByQueryAction.INSTANCE);
+            deleteByQuery.source(ContentIndex.INDEX_NAME).filter(QueryBuilders.matchAllQuery());
+
+            BulkByScrollResponse response = deleteByQuery.get();
+            log.debug(
+                    "[{}] wiped. {} documents were removed", ContentIndex.INDEX_NAME, response.getDeleted());
+        } catch (OpenSearchTimeoutException e) {
+            log.error("[{}] delete query timed out: {}", ContentIndex.INDEX_NAME, e.getMessage());
+        }
     }
 }